+++ CHANGELOG.md
@@ -26,6 +26,7 @@ _Note: Gaps between patch versions are faulty, broken or test releases._
 
 * Resolved an issue with the progress plugin's view in relation to empty builds
 * Corrected the theme-manager to function properly in server-side rendering (ssr)
+* Added synchronous rendering of the first chunk using `v-for`. This is because SSR does not have access to the DOM API required for `vdom`. Therefore, we leverage Vue functionality to render the first chunk equally for SSR and CSR. `components/base/b-virtual-scroll-new`
 
 ## v4.0.0-beta.76 (2023-03-25)